move profiles to component folder
