+++ src/components/Profile.tsx
@@ -45,7 +45,7 @@ function filterRoles(role: Role): boolean {
 
 type ProfileComponent = React.FC<ProfileProps>;
 
-export const Profile: ProfileComponent = ({ member, highestRole }) => {
+const Profile: ProfileComponent = ({ member, highestRole }) => {
   const avatarURL = `https://cdn.discordapp.com/avatars/${member.id}/${
     member.avatar
   }.${member.avatar.startsWith("a_") ? "gif" : "png"}?size=128`;
@@ -145,3 +145,64 @@ export const Profile: ProfileComponent = ({ member, highestRole }) => {
     </Card>
   );
 };
+
+export const Profiles: React.FC<{ role: Role }> = ({ role }) => (
+  <div
+    className={[
+      "grid",
+      "grid-cols-profile",
+      "gap-4",
+      "items-start",
+      "justify-center",
+    ].join(" ")}
+  >
+    {role.members.map((member) => (
+      <Profile
+        member={member}
+        highestRole={role}
+        key={`profile-${member.id}`}
+      />
+    ))}
+  </div>
+);
+
+export const descriptions: Record<string, string> = {
+  "568141992794783749":
+    "The first point of call for disputes between crew members. They monitor Discord and lobbies, and resolve conflicts within crew.",
+  "549644467498516508":
+    "Monitor Discord and lobbies, enforce the rules, and have the ability to ban/mute/kick. They are second in command to the Admin.",
+
+  "546342033867014165":
+    "The top of the hierarchy, they make the rules once agreed upon with the staff, maintain the Discord server, and the Reddit page. Crew members should come to them if they have any issues, feedback for staff members or anything crew related.",
+};
+
+const eventRole: Role = {
+  name: "Events",
+  color: "fake-events-role",
+  id: "fake-events-role",
+  position: 1,
+};
+
+const eventRoles = [
+  "609447212925190147", //events gta-xb
+  "608396947795083274", //events gta-ps
+  "760091219069763594", //events rdo-xb
+  "760091432114716692", //events rdo-ps
+];
+
+export const mappedStaffProfiles = (staffRoles: Role[]): Role[] =>
+  staffRoles.map((role) => ({
+    ...role,
+    members: role.members.map((member) => {
+      if (member.roles.find(({ id }) => eventRoles.includes(id))) {
+        return {
+          ...member,
+          roles: [
+            ...member.roles.filter(({ id }) => !eventRoles.includes(id)),
+            eventRole,
+          ],
+        };
+      }
+      return member;
+    }),
+  }));
